Extract NearbyLocationView
The plan is to extract UI from work for the ViewModifiers. This way the UI, the work, and the .toolbar parts can all be separated, in order to try out a single .toolbar for the StateView.
